python_agents/collector/gcp_provider.py
"""
Google Cloud Platform (GCP) implementation of cloud provider interfaces.
"""

import json
import logging
from datetime import datetime, timezone
from typing import Any, Dict, List, Optional

from .cloud_provider import (
    CloudConfig,
    CloudProvider,
    CloudProviderInterface,
    IAMCollectorInterface,
    SecurityCollectorInterface,
    LogCollectorInterface,
    CloudCollectorFactory
)

logger = logging.getLogger(__name__)


class GCPProvider(CloudProviderInterface):
    """Google Cloud Platform provider implementation."""
    
    def __init__(self, config: CloudConfig):
        super().__init__(config)
        self.iam_collector = GCPIAMCollector(config)
        self.security_collector = GCPSecurityCollector(config)
        self.log_collector = GCPLogCollector(config)
    
    @property
    def provider_name(self) -> str:
        return "Google Cloud Platform"
    
    def validate_credentials(self) -> bool:
        """Validate GCP credentials."""
        if self.config.use_mock:
            return True
        
        try:
            from google.auth import default
            credentials, project = default()
            return credentials is not None
        except Exception as e:
            logger.error(f"Failed to validate GCP credentials: {e}")
            return False


class GCPIAMCollector(IAMCollectorInterface):
    """GCP-specific IAM collector implementation."""
    
    def __init__(self, config: CloudConfig):
        self.config = config
        self.project_id = config.project_id or "example-project"
    
    def collect_users(self) -> List[Dict[str, Any]]:
        """Collect GCP user identities from IAM bindings."""
        if self.config.use_mock:
            return self._get_mock_users()
        
        # Real implementation would use google-cloud-iam
        logger.warning("Real GCP IAM user collection not implemented, using mock data")
        return self._get_mock_users()
    
    def collect_roles(self) -> List[Dict[str, Any]]:
        """Collect GCP IAM roles."""
        if self.config.use_mock:
            return self._get_mock_roles()
        
        # Real implementation would use google-cloud-iam
        logger.warning("Real GCP IAM role collection not implemented, using mock data")
        return self._get_mock_roles()
    
    def collect_policies(self) -> List[Dict[str, Any]]:
        """Collect GCP IAM policy bindings."""
        if self.config.use_mock:
            return self._get_mock_policies()
        
        try:
            from google.cloud import iam
            # Real implementation here
            logger.warning("Real GCP IAM policy collection not implemented, using mock data")
            return self._get_mock_policies()
        except ImportError:
            logger.error("google-cloud-iam not installed, using mock data")
            return self._get_mock_policies()
    
    def _get_mock_users(self) -> List[Dict[str, Any]]:
        return [
            {
                "type": "user",
                "email": "admin@example.com",
                "roles": ["roles/owner"],
                "provider": "gcp"
            },
            {
                "type": "user",
                "email": "developer@example.com",
                "roles": ["roles/owner", "roles/editor"],
                "provider": "gcp"
            },
            {
                "type": "serviceAccount",
                "email": "app-sa@project.iam.gserviceaccount.com",
                "roles": ["roles/editor"],
                "provider": "gcp"
            }
        ]
    
    def _get_mock_roles(self) -> List[Dict[str, Any]]:
        return [
            {
                "name": "roles/owner",
                "title": "Owner",
                "description": "Full access to all resources",
                "permissions": ["*"],
                "provider": "gcp"
            },
            {
                "name": "roles/editor",
                "title": "Editor",
                "description": "Edit access to all resources",
                "permissions": ["compute.*", "storage.*"],
                "provider": "gcp"
            }
        ]
    
    def _get_mock_policies(self) -> List[Dict[str, Any]]:
        return [
            {
                "resource": f"projects/{self.project_id}",
                "bindings": [
                    {
                        "role": "roles/owner",
                        "members": ["user:admin@example.com", "user:developer@example.com"]
                    },
                    {
                        "role": "roles/editor",
                        "members": ["serviceAccount:app-sa@project.iam.gserviceaccount.com"]
                    }
                ],
                "provider": "gcp"
            }
        ]


class GCPSecurityCollector(SecurityCollectorInterface):
    """GCP-specific security findings collector."""
    
    def __init__(self, config: CloudConfig):
        self.config = config
        self.organization_id = "123456"  # Default for mock
    
    def collect_findings(self, severity_filter: Optional[List[str]] = None) -> List[Dict[str, Any]]:
        """Collect Security Command Center findings."""
        if self.config.use_mock:
            findings = self._get_mock_findings()
        else:
            # Real implementation would use google-cloud-securitycenter
            logger.warning("Real GCP SCC collection not implemented, using mock data")
            findings = self._get_mock_findings()
        
        # Apply severity filter if provided
        if severity_filter:
            findings = [f for f in findings if f.get("severity") in severity_filter]
        
        return findings
    
    def collect_compliance_status(self) -> Dict[str, Any]:
        """Collect GCP compliance and security posture information."""
        if self.config.use_mock:
            return self._get_mock_compliance()
        
        # Real implementation would aggregate various compliance checks
        logger.warning("Real GCP compliance collection not implemented, using mock data")
        return self._get_mock_compliance()
    
    def _get_mock_findings(self) -> List[Dict[str, Any]]:
        return [
            {
                "id": "finding-1",
                "category": "OVERPRIVILEGED_SERVICE_ACCOUNT",
                "resource": "serviceAccounts/over-privileged-sa@example-project.iam.gserviceaccount.com",
                "severity": "HIGH",
                "status": "ACTIVE",
                "description": "Service account has excessive permissions",
                "provider": "gcp",
                "finding_type": "vulnerability"
            },
            {
                "id": "finding-2",
                "category": "PUBLIC_BUCKET",
                "resource": "storage.googleapis.com/example-public-bucket",
                "severity": "MEDIUM",
                "status": "ACTIVE",
                "description": "Storage bucket is publicly accessible",
                "provider": "gcp",
                "finding_type": "misconfiguration"
            }
        ]
    
    def _get_mock_compliance(self) -> Dict[str, Any]:
        return {
            "provider": "gcp",
            "standards": {
                "cis_gcp": {
                    "score": 0.75,
                    "passed": 150,
                    "failed": 50
                },
                "pci_dss": {
                    "score": 0.82,
                    "passed": 82,
                    "failed": 18
                }
            },
            "last_scan": datetime.now(timezone.utc).isoformat()
        }


class GCPLogCollector(LogCollectorInterface):
    """GCP-specific audit log collector."""
    
    def __init__(self, config: CloudConfig):
        self.config = config
        self.project_id = config.project_id or "example-project"
    
    def collect_recent_logs(self, hours: int = 24) -> List[Dict[str, Any]]:
        """Collect recent Cloud Audit Logs."""
        if self.config.use_mock:
            return self._get_mock_logs()
        
        # Real implementation would use google-cloud-logging
        logger.warning("Real GCP log collection not implemented, using mock data")
        return self._get_mock_logs()
    
    def collect_suspicious_activities(self) -> List[Dict[str, Any]]:
        """Collect logs indicating suspicious activities."""
        if self.config.use_mock:
            return self._get_mock_suspicious_activities()
        
        # Real implementation would query for specific patterns
        logger.warning("Real GCP suspicious activity collection not implemented, using mock data")
        return self._get_mock_suspicious_activities()
    
    def _get_mock_logs(self) -> List[Dict[str, Any]]:
        return [
            {
                "timestamp": datetime.now(timezone.utc).isoformat(),
                "severity": "INFO",
                "resource": {"type": "gce_instance", "labels": {"instance_id": "123456"}},
                "principal": "user:admin@example.com",
                "method": "compute.instances.delete",
                "status": "SUCCESS",
                "provider": "gcp"
            }
        ]
    
    def _get_mock_suspicious_activities(self) -> List[Dict[str, Any]]:
        return [
            {
                "timestamp": datetime.now(timezone.utc).isoformat(),
                "severity": "WARNING",
                "activity": "Multiple failed login attempts",
                "principal": "user:suspicious@example.com",
                "details": "10 failed login attempts in 5 minutes",
                "provider": "gcp"
            }
        ]


# Register GCP provider with the factory
CloudCollectorFactory.register(CloudProvider.GCP, GCPProvider)