python_agents/requirements.txt
# Google Cloud SDKs
google-cloud-iam>=2.0.0
google-cloud-securitycenter>=1.0.0
google-cloud-aiplatform>=1.0.0

# AWS SDKs
boto3>=1.26.0
botocore>=1.29.0

# Azure SDKs
azure-identity>=1.14.0
azure-mgmt-resource>=23.0.0
azure-mgmt-authorization>=4.0.0
azure-mgmt-security>=5.0.0
azure-mgmt-monitor>=6.0.0

# CLI and templating
fire>=0.5.0
jinja2>=3.0.0

# Testing and quality tools
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.0.0
black>=23.0.0
flake8>=6.0.0
coverage>=7.0.0

# Additional utilities
python-dateutil>=2.8.0